Continue cleaning up
